Update HomeController.cs
This file is the key for messing with the database. There are functions for retrieving and altering the username and subreddits from the database. For my tasks, I created the functions for retrieval/updating subreddits with the help of ai. These are the functions that will get called from all the views. Also, important functions under these are the callings for each of the subreddit pages. Zachary and I worked with the retrieval/alteration of the username functions; however, those are still in progress. Code review: Reviewed and approved by Zachary.